Add dotnet web app
