+++ examples/getting-started/GettingStarted.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApp", "WebApp.csproj", "{B93C924C-E830-48FC-AF67-3BCA0A5683E7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApp", "WebApp\WebApp.csproj", "{B8F33EE6-4840-4766-9A3B-2D73126F1E15}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -14,9 +14,9 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B93C924C-E830-48FC-AF67-3BCA0A5683E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B93C924C-E830-48FC-AF67-3BCA0A5683E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B93C924C-E830-48FC-AF67-3BCA0A5683E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B93C924C-E830-48FC-AF67-3BCA0A5683E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8F33EE6-4840-4766-9A3B-2D73126F1E15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8F33EE6-4840-4766-9A3B-2D73126F1E15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8F33EE6-4840-4766-9A3B-2D73126F1E15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8F33EE6-4840-4766-9A3B-2D73126F1E15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
