--- examples/getting-started/Pages/_ViewImports.cshtml
@@ -1,5 +1,0 @@
-﻿@using Phoria.Islands
-@using WebApp
-@namespace WebApp.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Phoria
