--- examples/getting-started/README.md
@@ -1,5 +1,0 @@
-# Getting started with Phoria example
-
-This example is the output of following the manual setup process described in the [Getting started with Phoria](https://github.com/CMeeg/phoria/blob/main/docs/guides/getting-started.md) guide.
-
-It is a basic example of adding Phoria Islands (using React) to a new Razor Pages web app created using the offical `webapp` template.
