+++ examples/getting-started/WebApp/Pages/Index.cshtml.cs
