+++ examples/getting-started/WebApp/Pages/Privacy.cshtml
