+++ examples/getting-started/WebApp/Pages/Shared/_ValidationScriptsPartial.cshtml
