+++ examples/getting-started/WebApp/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using WebApp
+@namespace WebApp.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
