+++ examples/getting-started/WebApp/Pages/_ViewStart.cshtml
