+++ examples/getting-started/WebApp/wwwroot/favicon.ico
