+++ examples/getting-started/WebApp/wwwroot/js/site.js
