+++ examples/getting-started/WebApp/wwwroot/lib/jquery-validation/LICENSE.md
