+++ examples/getting-started/WebApp/wwwroot/lib/jquery/LICENSE.txt
