--- examples/getting-started/appsettings.Development.json
@@ -1,14 +1,0 @@
-{
-  "detailedErrors": true,
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "phoria": {
-    "server": {
-      "https": true
-    }
-  }
-}
