--- examples/getting-started/appsettings.Preview.json
@@ -1,3 +1,0 @@
-{
-  "detailedErrors": true
-}
