--- examples/getting-started/appsettings.Production.json
@@ -1,10 +1,0 @@
-{
-  "phoria": {
-    "server": {
-      "process": {
-        "command": "node",
-        "arguments": ["ui/dist/server/server.js"]
-      }
-    }
-  }
-}
