--- examples/getting-started/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "allowedHosts": "*",
-  "phoria": {
-    "entry": "src/entry-client.ts",
-    "ssrEntry": "src/entry-server.ts"
-  }
-}
