--- examples/getting-started/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "getting-started",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "build": "run-p build:* -c",
-    "build:app": "vite build --app",
-    "build:dotnet": "dotnet build WebApp.csproj --configuration Release",
-    "build:server": "vite build --config vite.server.config.ts",
-    "dev": "tsx ./ui/src/server.ts",
-    "lint": "eslint .",
-    "preview": "run-p preview:* -c",
-    "preview:dotnet": "cross-env DOTNET_ENVIRONMENT=Preview dotnet run ./bin/Release/net9.0/WebApp.dll --no-launch-profile",
-    "preview:phoria": "cross-env NODE_ENV=production DOTNET_ENVIRONMENT=Preview node ./ui/dist/server/server.js"
-  },
-  "dependencies": {
-    "@phoria/phoria": "^0.4.1",
-    "@phoria/phoria-react": "^0.4.1",
-    "h3": "^1.13.0",
-    "listhen": "^1.9.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.17.0",
-    "@phoria/vite-plugin-dotnet-dev-certs": "^0.2.0",
-    "@types/node": "^22.10.5",
-    "@types/react": "^19.0.4",
-    "@types/react-dom": "^19.0.2",
-    "@vitejs/plugin-react": "^4.3.4",
-		"cross-env": "^7.0.3",
-    "eslint": "^9.17.0",
-    "eslint-plugin-react": "^7.37.3",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.14",
-    "globals": "^15.14.0",
-    "npm-run-all": "^4.1.5",
-    "tsx": "^4.19.2",
-    "typescript": "~5.7.3",
-    "typescript-eslint": "^8.19.1",
-    "vite": "^6.0.7"
-  },
-  "packageManager": "pnpm@9.15.0+sha256.09a8fe31a34fda706354680619f4002f4ccef6dadff93240d24ef6c831f0fd28"
-}
