--- examples/getting-started/ui/src/components/register.ts
@@ -1,8 +1,0 @@
-import { registerComponents } from "@phoria/phoria"
-
-registerComponents({
-  Counter: {
-    loader: () => import("./Counter/Counter.tsx"),
-    framework: "react"
-  }
-})
