--- examples/getting-started/ui/src/entry-client.ts
@@ -1,5 +1,0 @@
-import "@phoria/phoria-react/client"
-import "./components/register"
-import { PhoriaIsland } from "@phoria/phoria/client"
-
-PhoriaIsland.register()
