--- examples/getting-started/ui/src/entry-server.ts
@@ -1,9 +1,0 @@
-import "@phoria/phoria-react/server"
-import "./components/register"
-import type { PhoriaIsland } from "@phoria/phoria/server"
-
-async function renderPhoriaIsland(island: PhoriaIsland) {
-	return await island.render()
-}
-
-export { renderPhoriaIsland }
